+++ src/depyty/cli/__init__.py
@@ -6,7 +6,7 @@ This encompasses
 - A struct representing the CLI arguments entered by users
 - All subcommands and their logic in the `commands` module
 
-If you e.g. want to look what the `analyze` command does, go to
-`depyty.cli.commands.analyze` and inspect the `run` method. This gets called
-when a user runs `depyty analyze`.
+If you e.g. want to look what the `check` command does, go to
+`depyty.cli.commands.check` and inspect the `run` method. This gets called
+when a user runs `depyty check`.
 """